refactor: rename tests/mock_crust to tests/mock_network
