+++ tests/mock_network/accumulate.rs
