+++ tests/mock_network_tests.rs
@@ -117,4 +117,4 @@ macro_rules! expect_no_event {
     }};
 }
 
-mod mock_crust;
+mod mock_network;
